requirements.txt
pandas>=2.0
requests>=2.31
meteostat>=1.6
pyarrow>=15.0
pyyaml>=6.0
